modify header file

+++ sort.h
@@ -23,7 +23,7 @@ void bubble_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
 void quicksort(int *, int lo, int hi, size_t size);
 void _swap(int *, int p1, int p2, size_t size);
-size_t partition(int *, int lo, int hi, size_t size);
+int partition(int *, int lo, int hi, size_t size);
 
 #endif /* SORT_H */
